Minor code cleaning
